restructure react stories
