--- react/storybook-v5/src/stories/index.js
@@ -1,4 +1,0 @@
-// Stories
-import './button';
-import './toast';
-import './nav';
